+++ skvideo/measure/__init__.py
@@ -5,9 +5,13 @@
 from .ssim import *
 from .psnr import *
 from .mse import *
+from .scene import *
+from .edge import *
 
 __all__ = [
     'ssim',
     'mse',
     'psnr',
+    'scenedet',
+    'canny',
 ]